Move the slider to the right folder
