+++ src/components/modals/AccelerateTxModal.tsx
@@ -4,13 +4,13 @@ import { Text, View } from 'react-native'
 import { AirshipBridge } from 'react-native-airship'
 
 import s from '../../locales/strings'
-import { Slider } from '../../modules/UI/components/Slider/Slider'
 import { getDisplayDenominationFromState, getExchangeDenominationFromState } from '../../selectors/DenominationSelectors'
 import { connect } from '../../types/reactRedux'
 import { GuiExchangeRates } from '../../types/types'
 import { convertTransactionFeeToDisplayFee } from '../../util/utils'
 import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
 import { ModalFooter, ModalMessage, ModalTitle } from '../themed/ModalParts'
+import { Slider } from '../themed/Slider'
 import { ThemedModal } from '../themed/ThemedModal'
 import { Tile } from '../tiles/Tile'
