+++ src/components/scenes/FioConnectWalletConfirmScene.tsx
@@ -3,7 +3,6 @@ import { ScrollView } from 'react-native'
 
 import s from '../../locales/strings'
 import { FIO_NO_BUNDLED_ERR_CODE, updatePubAddressesForFioAddress } from '../../modules/FioAddress/util'
-import { Slider } from '../../modules/UI/components/Slider/Slider'
 import { CcWalletMap } from '../../reducers/FioReducer'
 import { connect } from '../../types/reactRedux'
 import { NavigationProp, RouteProp } from '../../types/routerTypes'
@@ -13,6 +12,7 @@ import { ButtonsModal } from '../modals/ButtonsModal'
 import { Airship, showError, showToast } from '../services/AirshipInstance'
 import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
 import { EdgeText } from '../themed/EdgeText'
+import { Slider } from '../themed/Slider'
 import { Radio } from '../themed/ThemedButtons'
 import { Tile } from '../tiles/Tile'
