+++ src/components/scenes/FioRequestConfirmationScene.tsx
@@ -6,7 +6,6 @@ import { View } from 'react-native'
 import { formatNumber } from '../../locales/intl'
 import s from '../../locales/strings'
 import { addToFioAddressCache, checkPubAddress, convertEdgeToFIOCodes, getRemainingBundles } from '../../modules/FioAddress/util'
-import { Slider } from '../../modules/UI/components/Slider/Slider'
 import { CcWalletMap } from '../../reducers/FioReducer'
 import { getDisplayDenomination, getExchangeDenomination } from '../../selectors/DenominationSelectors'
 import { getExchangeRate, getSelectedCurrencyWallet } from '../../selectors/WalletSelectors'
@@ -20,6 +19,7 @@ import { ButtonsModal } from '../modals/ButtonsModal'
 import { TextInputModal } from '../modals/TextInputModal'
 import { Airship, showError, showToast } from '../services/AirshipInstance'
 import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeContext'
+import { Slider } from '../themed/Slider'
 import { Tile } from '../tiles/Tile'
 
 interface StateProps {
