+++ src/components/scenes/FioStakingChangeScene.tsx
@@ -10,7 +10,6 @@ import fioLogo from '../../assets/images/fio/fio_logo.png'
 import { SPECIAL_CURRENCY_INFO, STAKING_BALANCES } from '../../constants/WalletAndCurrencyConstants'
 import { formatNumber, formatTimeDate, SHORT_DATE_FMT } from '../../locales/intl'
 import s from '../../locales/strings'
-import { Slider } from '../../modules/UI/components/Slider/Slider'
 import { getDisplayDenomination, getExchangeDenomination } from '../../selectors/DenominationSelectors'
 import { convertCurrency } from '../../selectors/WalletSelectors'
 import { connect } from '../../types/reactRedux'
@@ -24,6 +23,7 @@ import { cacheStyles, Theme, ThemeProps, withTheme } from '../services/ThemeCont
 import { EdgeText } from '../themed/EdgeText'
 import { ModalFooter, ModalMessage, ModalTitle } from '../themed/ModalParts'
 import { SceneHeader } from '../themed/SceneHeader'
+import { Slider } from '../themed/Slider'
 import { ThemedModal } from '../themed/ThemedModal'
 import { Tile } from '../tiles/Tile'
